UDP
finished UDP stuff with both int and string in the client and server. This also allows the user to type either a number or a string in the same application and if a number is typed it is multiplied by 2 or if a string is typed it is converted to uppercase and returned.